Added host files. Cleaned linux-advanced instalation

+++ instalation-scripts/linux-advanced.sh
@@ -1,17 +1,24 @@
 #!/bin/bash
 #Installs top.sls and advanced-linux-apps
+# Check if script has been run before
+if [ -f /tmp/helloworld.txt ]; then
+    echo "This install script has already been run! It is intended to be run only once."
+    exit 0
+fi
 cd
 echo "Updating packages..."
 sudo apt-get update -qq >> /dev/null
 echo "Installing git and salt..."
 sudo apt-get install git curl -y -qq >> /dev/null
 curl -L https://bootstrap.saltstack.com -o install_salt.sh
-sudo sh install_salt.sh -P
+sudo sh install_salt.sh -P -qq >> /dev/null
+sudo echo -e "master: localhost\nid: linux-advanced" | sudo tee /etc/salt/minion
 echo "Cloning repository..."
 cd /srv/
 sudo git clone https://github.com/aksratamo/salt
+cd salt/
+sudo wget https://someonewhocares.org/hosts/hosts
+
 
 #Running highstate and advanced-linux
 sudo salt-call --local state.apply
-sudo salt-call --local state.apply linux-apps
-sudo salt-call --local state.apply linux-advanced
